Switch debug builds to use clang.
I've often seen that we are unable to properly debug on customer machines
running rhel6 (e.g. "Can't compute CFA..."). This results in not having symbols
resolve correctly. A little poking around indicates these are bugs in GDB and
I think it will be pretty difficult to update GDB on our build machines. On
the other hand, they are all running a much newer version of clang which I hope
is less buggy/more likely to be patched if there are bugs/etc.

If people have a particular affinity for GCC for development, I can modify this
patch to respect that.

Change-Id: I4f655313384222732834cb3267c9954b6c825a57
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3279
Reviewed-by: Henry Robinson <henry@cloudera.com>
Tested-by: jenkins