Add flag to disable any memory pooling for debugging.
The things we do to reuse memory and minimize allocations reduces the
effectiveness of asan. This patch adds a startup flag (would only be
used for debugging) that disables the pooling. This means that allocations
are one to one with malloc/frees.

This found one bug in the pcsa UDA which conveniently proves its effectiveness.
Unfortunately, I doubt anyone is using that function.

Change-Id: Id13d22a5ebad4d31c298ccb237d83d5398799f39
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3301
Reviewed-by: Nong Li <nong@cloudera.com>
Tested-by: jenkins