Remove unnecessary decimal casts for some builtins.
For arithmetic ops, this is an optimization. The Add(decimal,decimal) already handles
the cast as part of the operation.

For binary predicates, the cast is bad and can lead to overflows. The decimal Compare()
function has custom logic to not overflow.

Change-Id: I9f5ad74ea89e9dfa5a3a40c1e07f7e9178bf1d52
(cherry picked from commit 6bffaa885542443ca559888d921853ecd194cbcb)
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3414
Reviewed-by: Skye Wanderman-Milne <skye@cloudera.com>
Tested-by: jenkins