Reading compressed text.
Adding the ability to read compressed text. Reading the compression type from the
file descriptors. Trying to homogenize a bit more the interface of the scanners.
Removing the LZO_TEXT file format, since it was not actually a file format.
Modifying the tests to load and test also text/{snap,gzip,bzip} databases.
Note that this patch requires some changes to Impala-lzo as well.

Change-Id: Ic0742ba11f106ba545050bdb71795efbff70ef74
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3549
Reviewed-by: Ippokratis Pandis <ipandis@cloudera.com>
Reviewed-by: Ishaan Joshi <ishaan@cloudera.com>
Tested-by: Ippokratis Pandis <ipandis@cloudera.com>