Updating the expectation of TestUnsupportedInsertFormats test.
Previously it was expecting that inserting in any combination of text/codec where
codec!=none would fail. With the read compressed text patch, such an insert will
succeed by inserting uncompressed text, ignoring the compression type. This is a
temporary fix for this (code coverage) test to succeed, until the text writers are
in.

Change-Id: Ib34bf661ee90e3bcbb76a225df8809ae13e835fa
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3659
Reviewed-by: Ippokratis Pandis <ipandis@cloudera.com>
Tested-by: jenkins