IMPALA-958: Long query plan serialization time when querying huge tables.
One easy thing to do, as mentioned by Ippo in the jira, is to avoid
sending the file_descs to the BE.  The BE doesn't use this info, but
instead reconstructs it from the scanranges.  When the blocks dominate
the TQueryExecRequest size, this reduces the size of serialized
TQueryExecRequest by around 30%.

I considered a couple of alternative approaches.  The first is to keep
sending the file_descs, but then make THdfsFileSplit reference them,
so that each scanrange doesn't carry the info that's common to a file.
The second would be to have the frontend just send the THdfsFileBlocks
rather than scanranges.  However, both of these get a little messy
because then more backend components need to know the difference
between Hdfs and HBase.

So, let's first see how much we can trim things down before making a
more drastic change.

Change-Id: Iecb63a22337839f3d6054046196ad323f690b64f
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3815
Reviewed-by: Daniel Hecht <dhecht@cloudera.com>
Tested-by: jenkins