IMPALA-1093: Catalog updates can fail with error: "IllegalArgumentException: fromKey out of range"
This bug was hit when garbage collecting the deleted catalog versions during a catalog
update on an impalad. The code that did the garbage collection assigned
log_ = log_.tailMap(), which returns a TreeMap backed by the orignal map but whose value
range is restricted by the values in the backing map. The correct thing to do is to just
create a new TreeMap so the value range is not restricted. This may also lead to a memory
leak since we never released a reference to the backing map.

No test included because I have not actually been able to reproduce this locally. Our DDL
stress test may hit this eventually.

Change-Id: I305ed01d33b0be6c8004717d0e6c0b4db6b4e606
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3741
Reviewed-by: Henry Robinson <henry@cloudera.com>
Tested-by: jenkins