IMPALA-1143: Tests don't succeed when just running buildall.sh.
There are at least two problems:

1) generate-schema-statements.py wasn't putting a newline on the very
last insert stmt, and beeline apparently was then ignoring it.

2) If HBaseTestDataRegionAssignment fails, then we reload a couple
tables, but were not recomputing stats for those tables.  And some
query-tests expect those tables to have stats.

Tesing: Ran the following commands and see that the tables are now
not-empty and include stats:

    $IMPALA_HOME/bin/load-data.py -w functional-query \
        --table_names=alltypesagg,alltypessmall --table_formats=hbase/none --force
    $IMPALA_HOME/tests/util/compute_table_stats.py --db_names=functional_hbase \
        --table_names=alltypesagg,alltypessmall

Change-Id: I5183e037d0f5499c81b79f2cc1060b71be2d4873
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3794
Reviewed-by: Nong Li <nong@cloudera.com>
Tested-by: jenkins