IMPALA-958: Don't store full paths in scan ranges.
When there are a lot of scan ranges per partition, we can reduce the
TQueryExecRequest footprint by storing only file names in the scan
ranges.  The directory path can be inferred by the partition id.

How much this reduces the footprint is dependent on the number of scan
ranges per parition and the length of the absolute path, but in our
test cluster for a partition with many files (~1000), this change
reduces the scan range size by about 30%.

Change-Id: I54e4e7290994c1c12975a75ceafd38fd20e6fbf6
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3844
Reviewed-by: Marcel Kornacker <marcel@cloudera.com>
Tested-by: jenkins
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4074
Tested-by: Daniel Hecht <dhecht@cloudera.com>