IMPALA-958: Don't store full paths in scan ranges.
When there are a lot of scan ranges per partition, we can reduce the
TQueryExecRequest footprint by storing only file names in the scan
ranges.  The directory path can be inferred by the partition id.

How much this reduces the footprint is dependent on the number of scan
ranges per parition and the length of the absolute path, but in our
test cluster for a partition with many files (~1000), this change
reduces the scan range size by about 30%.

Change-Id: I48fc72433dfd8ad7105c18694eae909c232fe8d7
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3844
Reviewed-by: Marcel Kornacker <marcel@cloudera.com>
Tested-by: jenkins