[CDH5] Re-enable JDBC tests by using Hive .12 JDBC driver
Re-enables the FE JDBC tests by using the Hive .12 JDBC driver. We need to be careful
that we don't mix Hive .12 and Hive .13 JARs outside of the test environment, so added
the dependency at the "test" scope and updated the dependency plugin to include
everything but "test" dependencies.
We actually invoke the mvn dependency plugin as part of mvn package, so I also removed
this call from buildall.

Change-Id: I0e92aab2ddbbf067421efa844192cd42409155a0
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3845
Reviewed-by: Lenni Kuff <lskuff@cloudera.com>
Tested-by: Lenni Kuff <lskuff@cloudera.com>