IMPALA-1142: Support specifying a custom AuthorizationProvider in Impala
This change allows Impala to support specifying a custom AuthorizationProvider class
(rather than the default HadoopGroupAuthorizationProvider). One use case is using the
LocalGroupAuthorizationProvider which makes it easier to test different users without
them actually existing and also allows Sentry be enabled without creating the groups
on all nodes the cluster. This is something we used to support but deprecated in
v1.4-cdh5 release. Added back in the changes and extended tests.

Change-Id: I09612e4e331689402f8f05d2666b3de61d881529
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3868
Tested-by: jenkins
Reviewed-by: Nong Li <nong@cloudera.com>
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3938
Reviewed-by: Lenni Kuff <lskuff@cloudera.com>