Fix code coverage build
We were testing casting 1/0 (inf) to integer types, which in C++ is
undefined. When codegen optimizations are enabled, this causes the
codegen'd cast to produce a different result than the non-codegen'd
cast. This patch removes these tests, i.e. Impala's behavior is
undefined as well. I'm open to other suggestions as to what we should
do in this case.

Change-Id: Ifdc11b5ec7c57e7c977ef9c41feb6db168c1a1d2
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3891
Reviewed-by: Nong Li <nong@cloudera.com>
Tested-by: jenkins