IMPALA-958: Prune unreferenced partitions from the descriptor table.
The planner prunes partitions before creating scan ranges, but we were still sending
all partition metadata from the FE back to the coordinator (and on to all BEs).  Let's
not bother sending metadata for partitions that are not referenced by any scan range
nor are part of a table that is the target of an insert.

Because the full descriptor table is sent to all backends involved, the savings will be
proportional to the number of backends (and the effectiveness of FE partition pruning).

Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3943
(cherry picked from commit f30eab417dc8425e4abb5668cc4fa6852526101d)
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3955

Change-Id: I823766aeb1bec261fb94d49c5dd3901492224ccb
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4076
Reviewed-by: Lenni Kuff <lskuff@cloudera.com>
Tested-by: Daniel Hecht <dhecht@cloudera.com>