IMPALA-958: Prune unreferenced partitions from the descriptor table.
The planner prunes partitions before creating scan ranges, but we were still sending
all partition metadata from the FE back to the coordinator (and on to all BEs).  Let's
not bother sending metadata for partitions that are not referenced by any scan range
nor are part of a table that is the target of an insert.

Because the full descriptor table is sent to all backends involved, the savings will be
proportional to the number of backends (and the effectiveness of FE partition pruning).

Change-Id: I6fee2c5957fd6d298361ac6582404487b045bd34
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3943
Reviewed-by: Daniel Hecht <dhecht@cloudera.com>
Tested-by: jenkins